+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.17.0] - 2025.06.19
+
+### Added
+
+- Added nonlinear dynamics in NetworkGraph
+
+### Fixed
+
+- Arrow direction for graph viz
+
 ## [5.16.0] - 2025.06.02
 
 ### Changed